move ImageAnimation into a component
